Add 3 components to FriendsPage in frontend
