+++ data/frontend/src/components/Achievements/Achievements.tsx
@@ -1,6 +1,6 @@
 import { Box, Grid, Button } from "@mui/material";
 import { Navigate, useNavigate, useParams } from "react-router-dom";
-import { useGetUserAchievementsQuery } from "../../../store/api";
+import { useGetUserAchievementsQuery } from "../../store/api";
 import Completed from "./Completed";
 import ToComplete from "./ToComplete";
